requirements.txt
fastapi
uvicorn[standard]
playwright
google-generativeai
websockets
pydantic
python-dotenv
hypothesis
pytest
pytest-asyncio
